requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
cycler==0.11.0
Flask==2.1.0
fonttools==4.31.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
torch
torchvision
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.0
python-telegram-bot
